[SPS-176] Auth 관련 의존성 API 모듈로 이전
