+++ clog-api/src/main/kotlin/org/depromeet/clog/server/api/auth/application/LogoutService.kt
@@ -1,11 +1,11 @@
-package org.depromeet.clog.server.domain.auth.application
+package org.depromeet.clog.server.api.auth.application
 
-import jakarta.transaction.Transactional
 import org.depromeet.clog.server.domain.auth.infrastructure.RefreshTokenRepository
 import org.depromeet.clog.server.domain.auth.presentation.exception.AuthException
 import org.depromeet.clog.server.domain.common.ErrorCode
 import org.depromeet.clog.server.domain.user.domain.UserRepository
 import org.springframework.stereotype.Service
+import org.springframework.transaction.annotation.Transactional
 
 @Service
 @Transactional
