+++ clog-api/src/main/kotlin/org/depromeet/clog/server/api/auth/application/dto/request/LocalLoginRequest.kt
@@ -1,0 +1,5 @@
+package org.depromeet.clog.server.api.auth.application.dto.request
+
+data class LocalLoginRequest(
+    val loginId: String
+)
