+++ clog-api/src/main/kotlin/org/depromeet/clog/server/api/auth/application/dto/request/RefreshTokenRequest.kt
@@ -1,0 +1,5 @@
+package org.depromeet.clog.server.api.auth.application.dto.request
+
+data class RefreshTokenRequest(
+    val refreshToken: String
+)
