+++ clog-api/src/main/kotlin/org/depromeet/clog/server/api/auth/application/strategy/AuthProviderHandler.kt
@@ -1,0 +1,7 @@
+package org.depromeet.clog.server.api.auth.application.strategy
+
+import org.depromeet.clog.server.api.auth.application.dto.response.AuthResponseDto
+
+interface AuthProviderHandler<T> {
+    fun login(request: T): AuthResponseDto
+}
