+++ clog-api/src/main/kotlin/org/depromeet/clog/server/api/auth/application/strategy/KakaoAuthProviderHandler.kt
@@ -1,16 +1,15 @@
-package org.depromeet.clog.server.domain.auth.application.strategy
+package org.depromeet.clog.server.api.auth.application.strategy
 
 import com.auth0.jwk.JwkProviderBuilder
 import com.auth0.jwt.JWT
 import com.auth0.jwt.JWTVerifier
 import com.auth0.jwt.algorithms.Algorithm
-import jakarta.transaction.Transactional
-import org.depromeet.clog.server.domain.auth.application.TokenService
-import org.depromeet.clog.server.domain.auth.application.dto.KakaoAccount
-import org.depromeet.clog.server.domain.auth.application.dto.KakaoProfile
-import org.depromeet.clog.server.domain.auth.application.dto.KakaoUserInfo
-import org.depromeet.clog.server.domain.auth.application.dto.request.KakaoLoginRequest
-import org.depromeet.clog.server.domain.auth.application.dto.response.AuthResponseDto
+import org.depromeet.clog.server.api.auth.application.TokenService
+import org.depromeet.clog.server.api.auth.application.dto.KakaoAccount
+import org.depromeet.clog.server.api.auth.application.dto.KakaoProfile
+import org.depromeet.clog.server.api.auth.application.dto.KakaoUserInfo
+import org.depromeet.clog.server.api.auth.application.dto.request.KakaoLoginRequest
+import org.depromeet.clog.server.api.auth.application.dto.response.AuthResponseDto
 import org.depromeet.clog.server.domain.auth.presentation.exception.AuthException
 import org.depromeet.clog.server.domain.common.ErrorCode
 import org.depromeet.clog.server.domain.user.domain.Provider
@@ -18,6 +17,7 @@ import org.depromeet.clog.server.domain.user.domain.User
 import org.depromeet.clog.server.domain.user.domain.UserRepository
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.stereotype.Service
+import org.springframework.transaction.annotation.Transactional
 import java.net.URL
 import java.security.interfaces.RSAPublicKey
 import java.util.concurrent.TimeUnit
