+++ clog-api/src/main/kotlin/org/depromeet/clog/server/api/user/presentation/UserController.kt
@@ -2,14 +2,14 @@ package org.depromeet.clog.server.api.user.presentation
 
 import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.tags.Tag
+import org.depromeet.clog.server.api.auth.application.LogoutService
 import org.depromeet.clog.server.api.configuration.ApiConstants.API_BASE_PATH_V1
 import org.depromeet.clog.server.api.configuration.annotation.ApiErrorCodes
 import org.depromeet.clog.server.api.user.UserContext
+import org.depromeet.clog.server.api.user.application.UserService
 import org.depromeet.clog.server.api.user.presentation.dto.WithdrawalRequest
-import org.depromeet.clog.server.domain.auth.application.LogoutService
 import org.depromeet.clog.server.domain.common.ClogApiResponse
 import org.depromeet.clog.server.domain.common.ErrorCode
-import org.depromeet.clog.server.domain.user.application.UserService
 import org.depromeet.clog.server.domain.user.application.dto.UpdateUserNameReauest
 import org.springframework.web.bind.annotation.*
