--- clog-domain/src/main/kotlin/org/depromeet/clog/server/domain/auth/application/dto/request/KakaoLoginRequest.kt
@@ -1,5 +1,0 @@
-package org.depromeet.clog.server.domain.auth.application.dto.request
-
-data class KakaoLoginRequest(
-    val idToken: String
-)
